third_party/blink/web_tests/wpt_internal/css/css-images/object-overflow-parsing.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS object-overflow: computed values</title>
<link rel="author" title="Khushal Sagar" href="mailto:khushalsagar@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<div id=target></div>
<script>
test_valid_value('object-overflow', 'clip');
test_valid_value('object-overflow', 'visible');
test_invalid_value('object-view-box', 'blah');

</script>
</body>
</html>
